// api/InlineObject12.ts
/* tslint:disable */
/* eslint-disable */
/**
 * ClassroomMgmtSys
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from './runtime';
/**
 *
 * @export
 * @interface InlineObject12
 */
export interface InlineObject12 {
    /**
     *
     * @type {number}
     * @memberof InlineObject12
     */
    id: number;
    /**
     *
     * @type {string}
     * @memberof InlineObject12
     */
    ip: string;
    /**
     *
     * @type {string}
     * @memberof InlineObject12
     */
    mac: string;
    /**
     *
     * @type {number}
     * @memberof InlineObject12
     */
    typeid: number;
    /**
     *
     * @type {number}
     * @memberof InlineObject12
     */
    classid: number;
}

export function InlineObject12FromJSON(json: any): InlineObject12 {
    return InlineObject12FromJSONTyped(json, false);
}

export function InlineObject12FromJSONTyped(json: any, ignoreDiscriminator: boolean): InlineObject12 {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {

        'id': json['id'],
        'ip': json['ip'],
        'mac': json['mac'],
        'typeid': json['typeid'],
        'classid': json['classid'],
    };
}

export function InlineObject12ToJSON(value?: InlineObject12 | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {

        'id': value.id,
        'ip': value.ip,
        'mac': value.mac,
        'typeid': value.typeid,
        'classid': value.classid,
    };
}